Adding support for strided input array
The input array is now an array containing separate inputs for each timestep with a stride the size of the number of neurons: `inputs[current_tick * neuron_count + sendr_id]`. This was added to support probabilistic inputs (poisson process) where applying the same `input` array 20x is no longer accurate.